% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euclide.R
\name{Euclide_R_for}
\alias{Euclide_R_for}
\title{Computing Euclidean Distance with a for loop}
\usage{
Euclide_R_for(v, w)
}
\arguments{
\item{v}{a vector of numeric data}

\item{w}{a vector of numeric data}
}
\value{
the Euclidean distance between v and w
}
\description{
Computing Euclidean Distance with a for loop
}
